move Grid out from create_surface_from_rectilinear_grid.py
